requirements.txt
torch
torchaudio
torchvision
openai
datasets
pandas
z3-solver
python-dotenv